% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{calc_timeline_limits}
\alias{calc_timeline_limits}
\title{Calculate the timeline limits}
\usage{
calc_timeline_limits(rfxstdt, rfxendt, rficdt = NULL, rfpendt = NULL)
}
\arguments{
\item{rfxstdt}{Treatment start date.}

\item{rfxendt}{Treatment end date.}

\item{rficdt}{Informed consent date (\code{NULL} value allowed).}

\item{rfpendt}{Participation end date  (\code{NULL} value allowed).}
}
\value{
Vector 2-tuple of timeline minimum and maximum limits.
}
\description{
Initialized to treatment start and end dates, but takes informed consent and participation end dates into account if
available. If end dates are missing then set to today's date.
}
\keyword{internal}
