% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_vlines}
\alias{create_vlines}
\title{Add vertical lines to plot}
\usage{
create_vlines(plot, plot_data, vline_vars, vline_day_numbers)
}
\arguments{
\item{plot}{A ggplot2 object}

\item{plot_data}{Data of the plot}

\item{vline_vars}{A list used to define the label and position of dashed lines in the plots}
}
\value{
A ggplot2 object
}
\description{
Add vertical lines to plot
}
\keyword{internal}
