% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palette.R
\name{fill_palette}
\alias{fill_palette}
\title{Create/complete a colour palette which maps colours to grading levels.}
\usage{
fill_palette(grading_vals, user_palette = NULL)
}
\arguments{
\item{grading_vals}{\verb{[character(1+)]} vector of grading values/levels.}

\item{user_palette}{\verb{[character(1+) | NULL]} named character vector mapping colour(s) to grading level(s).}
}
\value{
\verb{[character(1+)]} an updated colour palette where colours are mapped to ALL grading levels.
}
\description{
Create/complete a colour palette which maps colours to grading levels.
}
\keyword{internal}
