% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palette.R
\name{get_grading_vals}
\alias{get_grading_vals}
\title{Extract grading values using plot vars list and afmm data}
\usage{
get_grading_vals(input_plots, afmm_data)
}
\arguments{
\item{input_plots}{\verb{[list(1+)]}}

\item{afmm_data}{\verb{[list(1+)]}}
}
\value{
\verb{[character(1+)]}
}
\description{
Extract grading values using plot vars list and afmm data
}
\keyword{internal}
