% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_labels}
\alias{get_labels}
\title{This function will return corresponding labels in a dataset.}
\usage{
get_labels(dataset, col_names = NULL, keep_as_original = FALSE)
}
\arguments{
\item{dataset}{Dataset to get labels from.}

\item{col_names}{Vector of character names of the columns in the dataset.}

\item{keep_as_original}{Logical TRUE or FALSE: If TRUE, the corresponding column with
no label will return column name, else return NA.
If not assigned, it will return all labels of the the dataset. Otherwise,
only labels of  corresponding columns will be returned.}
}
\value{
Vector containing the labels of the dataset
}
\description{
This function will return corresponding labels in a dataset.
}
\keyword{internal}
