% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_patient_profile.R
\name{mock_patient_profile_app}
\alias{mock_patient_profile_app}
\title{Run papo mock}
\usage{
mock_patient_profile_app(auto_update_query_string = FALSE)
}
\description{
\code{mock_patient_profile_app()} runs the \pkg{dv.papo} module
with dummy data.
}
\keyword{internal}
