% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_patient_profile.R
\name{mock_patient_profile_server}
\alias{mock_patient_profile_server}
\title{Mock server for papo module}
\usage{
mock_patient_profile_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny.}
}
\description{
Uses adsl, adae, and adcm data of roche dataset.
}
\keyword{internal}
