% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_patient_profile.R
\name{mock_patient_profile_UI}
\alias{mock_patient_profile_UI}
\title{Mock UI for papo module}
\usage{
mock_patient_profile_UI(request)
}
\arguments{
\item{id}{A unique identifier.}
}
\value{
Shiny UI for the papomodule.
}
\description{
Mock UI for papo module
}
\keyword{internal}
