% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_patient_profile.R
\name{mod_patient_profile}
\alias{mod_patient_profile}
\title{Patient Profile Module}
\usage{
mod_patient_profile(
  module_id = "",
  subject_level_dataset_name = NULL,
  subjid_var = NULL,
  sender_ids = NULL,
  summary = NULL,
  listings = NULL,
  plots = NULL
)
}
\arguments{
\item{module_id}{\verb{[character]}
Unique Shiny module identifier.}

\item{subject_level_dataset_name}{\verb{[character]}
Subject-level dataset name.}

\item{subjid_var}{\verb{[character]}
Unique subject identifier column.}

\item{sender_ids}{\verb{[character(n)]} (optional)
Identifiers of modules allowed to request the display of target patient IDs.}

\item{summary}{\verb{[list]} (optional)
Subject-level summary section.
Composed of:
\itemize{
\item vars
\verb{[character(n)]}
Values to display in the summary section.
Indexes into dataset \code{subject_level_dataset_name}.
\item column_count
\verb{[integer]}
Column count for the summary section.
}}

\item{listings}{\verb{[list(n)]} (optional)
Listings section.
Composed of:
\itemize{
\item dataset
\verb{[character]}
Dataset name.
\item default_vars
\verb{[character(n)]} (optional)
Default columns to display. If not specified, the first six columns are
selected by default.
Indexes into dataset \code{dataset}.
}}

\item{plots}{\verb{[list]} (optional)
Plot section.
Composed of:
\itemize{
\item timeline_info
\verb{[list]}
Start and end study dates.
Composed of:
\itemize{
\item icf_date
\verb{[character]} (optional)
Informed Consent Form signing Date.
Indexes into dataset \code{subject_level_dataset_name}. Expects \verb{[Date|POSIXt]}
values.
\item trt_start_date
\verb{[character]}
Treatment Start Date, used also as Day 1 Reference Date.
Indexes into dataset \code{subject_level_dataset_name}. Expects \verb{[Date|POSIXt]}
values.
\item trt_end_date
\verb{[character]}
Treatment End Date.
Indexes into dataset \code{subject_level_dataset_name}. Expects \verb{[Date|POSIXt]}
values.
\item part_end_date
\verb{[character]} (optional)
Participation End Date.
Indexes into dataset \code{subject_level_dataset_name}. Expects \verb{[Date|POSIXt]}
values.
}
\item range_plots
\verb{[list(n)]}
Plots for range-like events (adverse events, concomitant medications, ...).
Composed of:
\itemize{
\item dataset
\verb{[character]}
Dataset name.
\item vars
\verb{[list]}
Variables of interest for the event.
Composed of:
\itemize{
\item start_date
\verb{[character]}
Start date.
Indexes into dataset \code{dataset}. Expects \verb{[Date|POSIXt]} values.
\item end_date
\verb{[character]}
End date.
Indexes into dataset \code{dataset}. Expects \verb{[Date|POSIXt]} values.
\item decode
\verb{[character]}
Descriptive event label.
Indexes into dataset \code{dataset}. Expects \verb{[character|factor]} values.
\item grading
\verb{[character]} (optional)
Degree of the event, used for color-coding the plot.
Indexes into dataset \code{dataset}. Expects \verb{[character|factor]} values.
\item serious_ae
\verb{[character]} (optional)
Marker of event seriousness, usually reserved for adverse event datasets.
Indexes into dataset \code{dataset}. Expects \verb{[logical|"Y"/"N"]} values.
}
\item tooltip
\verb{[character(n)]}
Block of text to display as hover information over the left edge of each range.
The names of this list are included as literal text and honor three basic HTML
formatting elements: \verb{<b>}, \verb{<i>}, \verb{<br>}). The columns the values refer to are
populated with the value on the dataset relevant to any given row.
Indexes into dataset \code{dataset}.
}
\item value_plots
\verb{[list(n)]}
Plots for value-like (lab measurements, vital signs, ...) traces.
Composed of:
\itemize{
\item dataset
\verb{[character]}
Dataset name.
\item vars
\verb{[list]}
Variables of interest for the trace.
Composed of:
\itemize{
\item analysis_param
\verb{[character]}
Parameter.
Indexes into dataset \code{dataset}. Expects \verb{[character|factor]} values.
\item analysis_val
\verb{[character]}
Value.
Indexes into dataset \code{dataset}. Expects \verb{[numeric]} values.
\item analysis_date
\verb{[character]}
Date.
Indexes into dataset \code{dataset}. Expects \verb{[Date|POSIXt]} values.
\item analysis_indicator
\verb{[character]} (optional)
Analysis reference range indicator.
Indexes into dataset \code{dataset}. Expects \verb{[character|factor]} values.
\item range_low_limit
\verb{[character]} (optional)
Lower limit of the reference range.
Indexes into dataset \code{dataset}. Expects \verb{[numeric]} values.
\item range_high_limit
\verb{[character]} (optional)
Upper limit of the reference range.
Indexes into dataset \code{dataset}. Expects \verb{[numeric]} values.
\item summary_stats
\verb{[character]} (optional)
Additional value column for summary statistics.
Indexes into dataset \code{dataset}. Expects \verb{[numeric]} values.
}
\item tooltip
\verb{[character(n)]}
Block of text to display as hover information over each point of the trace. The
names of this list are included as literal text and honor three basic HTML
formatting elements: \verb{<b>}, \verb{<i>}, \verb{<br>}). The columns the values refer to are
populated with the value on the dataset relevant to any given row.
Indexes into dataset \code{dataset}.
}
\item vline_vars
\verb{[character(n)]}
Place vertical dashed lines on days indicated by this dataset columns.
Indexes into dataset \code{subject_level_dataset_name}. Expects
\verb{[integer|Date|POSIXt]} values.
\item vline_day_numbers
\verb{[integer(n)]} (optional)
Place vertical dashed lines on days indicated by this parameter.
Represents a CDISC (non-zero) Study Day.
\item palette
\verb{[character(n)]} (optional)
If a name on this list matches the text on a plot element, the associated color
will be applied to that element. This mapping takes precedence over the
built-in palette.
Contains either an HTML (#xxxxxx) or an R color.
}}
}
\value{
A list composed of the following elements:
\itemize{
\item{\code{ui}}: Shiny UI function.
\item{\code{server}}: Shiny server function.
\item{\code{module_id}}: Shiny unique identifier.
}
}
\description{
\code{mod_patient_profile} is a DaVinci Shiny module that displays patient information on a subject-level basis. It
consists of three sections: summary, listings and plots.
\itemize{
\item The \code{summary} section shows visit-independent demographic information.
\item The \code{listings} section offers listings for visit-dependent information.
\item The \code{plots} section displays charts for events that happen over a span of time (adverse events, concomitant
medications, ...) or line plots of point-like events (laboratory values, vital sign measurements).
}\if{html}{\out{
<details>
<summary>Show/hide usage example </summary>
<div>
<br>
What follows is a sample call to the <span style='font-family:monospace'>mod_patient_profile</span> function.
See the <a href="../articles/a00-papo.html">main article of this package</a> for a tutorial on how to parameterize it.
<pre>
dv.papo::mod_patient_profile(
  module_id = "papo",
  subject_level_dataset_name = "adsl",
  subjid_var = "USUBJID",
  summary = list(
    vars = c("SUBJID", "SITEID", "ARM", "TRTSDT", "TRTEDT", "AGE", "RACE", "SEX", "BMIBL"),
    column_count = 3L
  ),
  listings = list(
    "Concomitant Medication" = list(
      dataset = "cm"
    ),
    "Adverse Events" = list(
      dataset = "adae",
      default_vars = c("ASTDT", "ASTDY", "AENDT", "AENDY", "AEDECOD", "AESEV")
    )
  ),
  plots = list(
    timeline_info = c(trt_start_date = "TRTSDT", trt_end_date = "TRTEDT"),
    range_plots = list(
      "Concomitant Medication" = list(
        dataset = "cm",
        vars = c(
          start_date = "CMSTDT", end_date = "CMENDT",
          decode = "CMDECOD", grading = "CMDECOD"
        ),
        tooltip = c()
      ),
      "Adverse Events" = list(
        dataset = "adae",
        vars = c(
          start_date = "ASTDT", end_date = "AENDT", decode = "AEDECOD",
          grading = "AESEV", serious_ae = "AESER"
        ),
        tooltip = c("AE Start Day: " = "ASTDY", "AE End Day: " = "AENDY")
      )
    ),
    value_plots = list(
      "Lab Values" = list(
        dataset = "lb",
        vars = c(
          analysis_param = "PARAM", analysis_val = "AVAL", analysis_date = "ADT",
          analysis_indicator = "LBNRIND", range_low_limit = "A1LO", range_high_limit = "A1HI"
        ),
        tooltip = c()
      )
    ),
    vline_vars = c(
      "Informed Consent Date" = "RFICDT"
    )
  )
)
</pre>
</div>
</details>
<hr>
}}
}
\seealso{
\code{\link[=mod_patient_profile_UI]{mod_patient_profile_UI()}}, \code{\link[=mod_patient_profile_server]{mod_patient_profile_server()}}
}
