% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_patient_profile.R
\name{mod_patient_profile_UI}
\alias{mod_patient_profile_UI}
\title{Patient Profile Module UI}
\usage{
mod_patient_profile_UI(id)
}
\arguments{
\item{id}{\verb{[character]} Unique shiny ID. Must match the ID provided to \code{\link[=mod_patient_profile_server]{mod_patient_profile_server()}}.}
}
\value{
Shiny UI.
}
\description{
(For use outside of the DaVinci framework)\cr
Places the Patient Profile module UI at the call site of this function. A matching call to \code{\link[=mod_patient_profile_server]{mod_patient_profile_server()}}
is necessary.\cr
}
\seealso{
\code{\link[=mod_patient_profile]{mod_patient_profile()}} and \code{\link[=mod_patient_profile_server]{mod_patient_profile_server()}}
}
