% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_patient_info.R
\name{patient_info_server}
\alias{patient_info_server}
\title{Create server for the patient information shiny module of \pkg{dv.papo}}
\usage{
patient_info_server(id, record, subjid_var, column_count = 3)
}
\arguments{
\item{id}{A unique ID string to create a namespace. Must match the ID of
\code{patient_info_UI()}.}

\item{record}{row object of dataframe: row record for patient.}

\item{subjid_var}{\verb{[character]}
Unique subject identifier column.}

\item{column_count}{A number indicating how many items should be in one row}
}
\description{
Create server for the patient information shiny module of \pkg{dv.papo}
}
\keyword{internal}
