% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_patient_info.R
\name{patient_info_UI}
\alias{patient_info_UI}
\title{Create user interface for the patient information shiny module of \pkg{dv.papo}}
\usage{
patient_info_UI(id)
}
\arguments{
\item{id}{A unique ID string to create a namespace. Must match the ID of
\code{patient_info_server()}.}
}
\value{
A shiny \code{uiOutput} element.
}
\description{
Create user interface for the patient information shiny module of \pkg{dv.papo}
}
\keyword{internal}
