% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_listing.R
\name{patient_listing_UI}
\alias{patient_listing_UI}
\title{Create user interface for patient listings shiny module of \pkg{dv.papo}}
\usage{
patient_listing_UI(id)
}
\arguments{
\item{id}{A unique ID string to create a namespace. Must match the ID of
\code{patient_plot_server()}.}
}
\description{
Create user interface for patient listings shiny module of \pkg{dv.papo}
}
\keyword{internal}
