% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{patient_plot_UI}
\alias{patient_plot_UI}
\title{Create user interface for patient plot shiny module of \pkg{dv.papo}}
\usage{
patient_plot_UI(id)
}
\arguments{
\item{id}{A unique ID string to create a namespace. Must match the ID of
\code{patient_plot_server()}.}

\item{title}{character: Title of plot module}
}
\description{
Create user interface for patient plot shiny module of \pkg{dv.papo}
}
\keyword{internal}
