% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{pt_info_data_filter}
\alias{pt_info_data_filter}
\title{Filter datasets for patient information section}
\usage{
pt_info_data_filter(df, subjid_var, columns, selected_key)
}
\arguments{
\item{df}{TODO}

\item{subjid_var}{\verb{[character]}
Unique subject identifier column.}

\item{columns}{TODO}

\item{selected_key}{TODO}
}
\value{
A list containing datasets for patient information and plots
}
\description{
Filter datasets for patient information section
}
\keyword{internal}
