% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{robust_ymd}
\alias{robust_ymd}
\title{Convert possibly truncated character(n) 'yyyy-mm-dd' to Date(n)
performing optional round up using the level of precision present
in the input data}
\usage{
robust_ymd(data, round_up = FALSE)
}
\arguments{
\item{data}{\link{character(n)} Vector of dates}
}
\description{
Convert possibly truncated character(n) 'yyyy-mm-dd' to Date(n)
performing optional round up using the level of precision present
in the input data
}
\keyword{internal}
