% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_ae_cm_plot}
\alias{create_ae_cm_plot}
\title{Function to create ae or cm plot}
\usage{
create_ae_cm_plot(
  data,
  x_limits,
  palette,
  sl_info,
  vline_vars,
  vline_day_numbers,
  x_axis_unit,
  x_axis_breaks,
  ref_date
)
}
\arguments{
\item{data}{Data frame containing the data for the plot}

\item{palette}{Named vector that contains the colors that are used in the plot}

\item{limits}{Vector that contains the limits of the plot}
}
\value{
A ggplot2 object
}
\description{
Function to create ae or cm plot
}
\keyword{internal}
