% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_lb_vs_plot}
\alias{create_lb_vs_plot}
\title{Function to create lab or vital sign plot}
\usage{
create_lb_vs_plot(
  data,
  date,
  val,
  low_limit,
  high_limit,
  param_var,
  param_val,
  summary_stats,
  x_limits,
  x_axis_unit,
  x_axis_breaks,
  palette,
  sl_info,
  vline_vars,
  vline_day_numbers,
  ref_date,
  plot_name,
  annotate_x_axis
)
}
\arguments{
\item{data}{Data frame containing the data for the plot}

\item{val}{Name of the variable that contains analyze value}

\item{low_limit}{Name of the variable that contains the values of the low limit of the normal range}

\item{high_limit}{Name of the variable that contains the values of the high limit of the normal range}

\item{param_var}{Name of the variable that contains the analysis parameter values}

\item{param_val}{Name of the analysis parameter}

\item{summary_stats}{Name of the variable that contains the values of the summary statistic}

\item{plot_name}{Name of plot}

\item{annotate_x_axis}{Logical indicating whether to annotate the x-axis}

\item{lb_selected_params}{Vector containing the values of the selected parameters}

\item{day}{Name of the variable that contains the analyze day}

\item{analysis_indicator}{Name of the variable that contains the values analysis indicator}

\item{limits}{Vector that contains the limits of the plot}
}
\value{
A ggplot2 object
}
\description{
Function to create lab or vital sign plot
}
\keyword{internal}
