% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_lb_vs_plot}
\alias{create_lb_vs_plot}
\title{Function to create lab or vital sign plot}
\usage{
create_lb_vs_plot(
  data,
  date,
  val,
  low_limit,
  high_limit,
  param,
  summary_stats,
  x_limits,
  x_axis_unit,
  x_axis_breaks,
  palette,
  sl_info,
  vline_vars,
  vline_day_numbers,
  ref_date
)
}
\arguments{
\item{data}{Data frame containing the data for the plot}

\item{val}{Character name of the variable that contains analyze value}

\item{low_limit}{Character name of the variable that contains the values of the low limit of the normal range}

\item{high_limit}{Character name of the variable that contains the values of the high limit of the normal range}

\item{summary_stats}{Character name of the variable that contains the values of the summary statistic}

\item{lb_selected_params}{Vector containing the values of the selected parameters}

\item{day}{Character name of the variable that contains the analyze day}

\item{params}{Character name of the variable that contains the values of the analyze parameters}

\item{analysis_indicator}{Character name of the variable that contains the values analysis indicator}

\item{limits}{Vector that contains the limits of the plot}
}
\value{
A ggplot2 object
}
\description{
Function to create lab or vital sign plot
}
\keyword{internal}
