% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dressing_room.R
\name{explorer_app}
\alias{explorer_app}
\title{Interactive demo/configuration tool for \code{\link[=mod_patient_profile]{mod_patient_profile()}}}
\usage{
explorer_app(datasets = NULL)
}
\arguments{
\item{datasets}{\verb{[list(data.frame(n))]} (optional) Datasets available to the module. One of them should be a
demographic subject-level dataset and the rest should be visit-dependent datasets. If not provided, the UI offers a
file input selector that is functionally equivalent.}
}
\description{
Launch an interactive configuration app for \code{\link[=mod_patient_profile]{mod_patient_profile()}}. This application guides the user through the
configuration of the module, offering an experimental point-and-click interface to the module API. Help is accessible
by hovering over any of the provided parameters. \cr
To try it using demo data, run \code{dv.papo::explorer_app(dv.papo:::prep_safety_data())} in your R prompt.
}
