% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palette.R
\name{fill_palette}
\alias{fill_palette}
\title{Create/complete a colour palette which maps colours to grading levels.}
\usage{
fill_palette(input_data, grading_cols, existing_palette = NULL)
}
\arguments{
\item{input_data}{\verb{[data.frame]} data.frame containing data of interest.}

\item{grading_cols}{\verb{[character(1+)]} vector of grading column names.}

\item{existing_palette}{\verb{[character(1+) | NULL]} named character vector mapping colour(s) to grading level(s).}
}
\value{
\verb{[character(1+)]} an updated colour palette where colours are mapped to ALL grading levels.
}
\description{
Create/complete a colour palette which maps colours to grading levels.
}
\keyword{internal}
