% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_listing.R
\name{patient_listing_server}
\alias{patient_listing_server}
\title{Create server for patient listings shiny module of \pkg{dv.papo}}
\usage{
patient_listing_server(id, data_list, key_value, listings)
}
\arguments{
\item{id}{A unique ID string to create a namespace. Must match the ID of
\code{patient_listing_UI()}.}

\item{data_list}{List of data frames containing data for each listing of selected patient.}

\item{key_value}{Character: Value of selected patient}

\item{listings}{\verb{[list(n)]} (optional)
Listings section.
Composed of:
\itemize{
\item dataset
\verb{[character]}
Dataset name.
\item default_vars
\verb{[character(n)]} (optional)
Default columns to display. If not specified, the first six columns are
selected by default.
Indexes into dataset \code{dataset}.
}}
}
\description{
Create server for patient listings shiny module of \pkg{dv.papo}
}
\keyword{internal}
