% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{patient_plot_server}
\alias{patient_plot_server}
\title{Create server for patient plot shiny module of \pkg{dv.papo}}
\usage{
patient_plot_server(
  id,
  subject_var,
  subject_level_dataset,
  timeline_info,
  extra_datasets,
  range_plots,
  value_plots,
  vline_vars,
  vline_day_numbers,
  palette
)
}
\description{
TODO: Document params (inherit them)
}
\keyword{internal}
