% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_safety_data.R
\name{prep_safety_data}
\alias{prep_safety_data}
\title{Prepare safety data}
\usage{
prep_safety_data(n = 200)
}
\arguments{
\item{n}{Number of rows to select from the adsl dataset. The first n rows will be
taken. Used to reduce runtime during development.}
}
\value{
A list of three data frames.
}
\description{
Modifiy safetyData's adsl, adae, and adcm dummy data for easy use within
\pkg{dv.clinlines}.
}
\keyword{internal}
